Give ability to run supervisor for dev
This will allow us to run supervisord in www-dev
that will run the site and a process to keep l10n
updated. It allows us to run this without also running
the other cron tasks that update the DB in every container,
since we use a separate clock container for that. Demo instances
will also run the DB update process via supervisor.
