+++ bedrock/scheduler_utils.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-
 from __future__ import print_function, unicode_literals
 
 import datetime
@@ -8,15 +6,10 @@ import sys
 from subprocess import check_call
 
 import requests
-from apscheduler.schedulers.blocking import BlockingScheduler
 from decouple import config
 from pathlib import Path
 
-
-schedule = BlockingScheduler()
 DEAD_MANS_SNITCH_URL = config('DEAD_MANS_SNITCH_URL', default='')
-DEV = config('DEV', cast=bool, default=False)
-
 # ROOT path of the project. A pathlib.Path object.
 ROOT_PATH = Path(__file__).resolve().parents[1]
 ROOT = str(ROOT_PATH)
@@ -30,14 +23,15 @@ def call_command(command):
 class scheduled_job(object):
     """Decorator for scheduled jobs. Takes same args as apscheduler.schedule_job."""
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, schedule, *args, **kwargs):
+        self.schedule = schedule
         self.args = args
         self.kwargs = kwargs
 
     def __call__(self, fn):
         self.name = fn.__name__
         self.callback = fn
-        schedule.add_job(self.run, id=self.name, *self.args, **self.kwargs)
+        self.schedule.add_job(self.run, id=self.name, *self.args, **self.kwargs)
         self.log('Registered')
         return self.run
 
@@ -70,54 +64,3 @@ def ping_dms(function):
 
     _ping.__name__ = function.__name__
     return _ping
-
-
-@scheduled_job('interval', minutes=30)
-@ping_dms
-def update_product_details():
-    call_command('update_product_details_files --database bedrock')
-
-
-@scheduled_job('interval', minutes=30)
-def update_externalfiles():
-    call_command('update_externalfiles')
-
-
-@scheduled_job('interval', minutes=30)
-def update_security_advisories():
-    call_command('update_security_advisories')
-
-
-@scheduled_job('interval', minutes=5)
-def rnasync():
-    # running in a subprocess as rnasync was not designed for long-running process
-    call_command('rnasync')
-
-
-@scheduled_job('interval', hours=6)
-def update_tweets():
-    call_command('cron update_tweets')
-
-
-@scheduled_job('interval', hours=1)
-def ical_feeds():
-    call_command('cron update_ical_feeds')
-    call_command('cron cleanup_ical_events')
-
-
-@scheduled_job('interval', hours=1)
-def update_firefox_os_feeds():
-    call_command('runscript update_firefox_os_feeds')
-
-
-if DEV:
-    @scheduled_job('interval', minutes=10)
-    def update_locales():
-        call_command('l10n_update')
-
-
-if __name__ == '__main__':
-    try:
-        schedule.start()
-    except (KeyboardInterrupt, SystemExit):
-        pass
