+++ etc/supervisor_available/cron_db.conf
@@ -1,0 +1,8 @@
+[program:cron_db]
+command = /app/bin/cron_db.py
+numprocs = 1
+autostart = true
+user = webdev
+redirect_stderr = true
+stdout_logfile = /dev/stdout
+stdout_logfile_maxbytes = 0
