+++ etc/supervisor_available/cron_l10n.conf
@@ -1,5 +1,5 @@
-[program:cron]
-command = /app/bin/cron.py
+[program:cron_l10n]
+command = /app/bin/cron_l10n.py
 numprocs = 1
 autostart = true
 user = webdev
